+++ CHANGELOG.md
@@ -6,6 +6,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.0.4] - 2021-04-26
+### Added
+- Scan() implemented
+- ScanWithContext() implemented
+- ScanPages() implemented
+- ScanPagesWithContext() implemented
+
 ## [0.0.3] - 2021-04-07
 ### Added
 - QueryPages() implemented